requirements.txt
alembic==0.9.2
click==6.7
ecdsa==0.13
Flask==0.12.2
Flask-Login==0.4.0
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
future==0.16.0
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
pycrypto==2.6.1
PyJWT==1.5.0
python-dateutil==2.6.0
python-editor==1.0.3
python-jose==1.3.2
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.10
Werkzeug==0.12.2
